pyproject.toml
[tool.poetry]
name = "bark-guard"
version = "0.1.0"
description = "The project will recognize barking and makes action"
authors = ["Lis Przemyslaw <przemyslaw_lis@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "*"
sounddevice = "*"
wavio = "*"
fastapi = "*"
uvicorn = "*"
sqlalchemy = "*"
PyJWT = "*"
matplotlib = "*"
librosa = "0.10.2"
llvmlite = "0.43.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
ruff = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
